# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KidWell Children's Health App API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProfilePetMood =
  (typeof ProfilePetMood)[keyof typeof ProfilePetMood];

export const ProfilePetMood = {
  happy: "happy",
  neutral: "neutral",
  sad: "sad",
} as const;

export interface Profile {
  id: string;
  name: string;
  age: number;
  avatarColor: string;
  petName: string;
  petLevel: number;
  petMood: ProfilePetMood;
  streakDays: number;
  totalPoints: number;
  createdAt: string;
}

export interface CreateProfileRequest {
  name: string;
  age: number;
  avatarColor: string;
  petName: string;
}

export interface Goals {
  fruitsVeggies: number;
  activityMinutes: number;
  sleepHours: number;
  screenTimeHours: number;
  waterGlasses: number;
}

export interface Dashboard {
  profileId: string;
  date: string;
  fruitsVeggies: number;
  activityMinutes: number;
  sleepHours: number;
  screenTimeHours: number;
  waterGlasses: number;
  goals: Goals;
  score: number;
}

export type CreateLogRequestType =
  (typeof CreateLogRequestType)[keyof typeof CreateLogRequestType];

export const CreateLogRequestType = {
  nutrition: "nutrition",
  activity: "activity",
  sleep: "sleep",
  screen_time: "screen_time",
  water: "water",
} as const;

export interface CreateLogRequest {
  profileId: string;
  type: CreateLogRequestType;
  value: number;
  notes?: string;
  date?: string;
}

export interface ActivityLog {
  id: string;
  profileId: string;
  type: string;
  value: number;
  notes?: string;
  date: string;
  createdAt: string;
}

export interface Achievement {
  id: string;
  profileId: string;
  type: string;
  title: string;
  description: string;
  icon: string;
  earnedAt: string;
  points: number;
}

export type GetLogsParams = {
  profileId: string;
  date?: string;
};