# lib/api-zod/src/generated/types/achievement.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KidWell Children's Health App API
 * OpenAPI spec version: 0.1.0
 */

export interface Achievement {
  id: string;
  profileId: string;
  type: string;
  title: string;
  description: string;
  icon: string;
  earnedAt: string;
  points: number;
}