# lib/api-zod/src/generated/types/activityLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KidWell Children's Health App API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityLog {
  id: string;
  profileId: string;
  type: string;
  value: number;
  notes?: string;
  date: string;
  createdAt: string;
}