# lib/api-zod/src/generated/types/createLogRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KidWell Children's Health App API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateLogRequestType } from "./createLogRequestType";

export interface CreateLogRequest {
  profileId: string;
  type: CreateLogRequestType;
  value: number;
  notes?: string;
  date?: string;
}