# lib/api-zod/src/generated/types/createProfileRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KidWell Children's Health App API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProfileRequest {
  name: string;
  age: number;
  avatarColor: string;
  petName: string;
}