# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KidWell Children's Health App API
 * OpenAPI spec version: 0.1.0
 */
import type { Goals } from "./goals";

export interface Dashboard {
  profileId: string;
  date: string;
  fruitsVeggies: number;
  activityMinutes: number;
  sleepHours: number;
  screenTimeHours: number;
  waterGlasses: number;
  goals: Goals;
  score: number;
}