# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KidWell Children's Health App API
 * OpenAPI spec version: 0.1.0
 */
import type { ProfilePetMood } from "./profilePetMood";

export interface Profile {
  id: string;
  name: string;
  age: number;
  avatarColor: string;
  petName: string;
  petLevel: number;
  petMood: ProfilePetMood;
  streakDays: number;
  totalPoints: number;
  createdAt: string;
}